requirements.txt
fastapi
sqlalchemy
psycopg2-binary
passlib[bcrypt]
pydantic
PyJWT
uvicorn
